Final posted version
Finished the end game loop with a UI pop up that congratulates the player with a resume or restart button.

Player will now not only be hurt but be respawned after they die, players will be respawned back to the first room.

Cleaned up code and added a few more overlays for interactions.